engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opendraft"
version = "1.6.22"
description = "Generate master-level research papers with real citations in minutes"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Federico De Ponte"}
]
maintainers = [
    {name = "Federico De Ponte"}
]
keywords = [
    "academic-writing",
    "draft",
    "ai",
    "research",
    "citation-management",
    "llm",
    "pdf-generation",
    "scholarly"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.20.0",
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",
    "pybtex>=0.24.0",
    "citeproc-py>=0.6.0",
    "PyYAML>=6.0.0",
    "markdown>=3.5.0",
    "weasyprint>=60.0",
    "python-docx>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]
docker = [
    "streamlit>=1.28.0",
]
all = [
    "opendraft[dev,docker]",
]

[project.urls]
Homepage = "https://opendraft.xyz"
Documentation = "https://github.com/federicodeponte/opendraft/tree/main/docs"
Repository = "https://github.com/federicodeponte/opendraft"
Changelog = "https://github.com/federicodeponte/opendraft/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/federicodeponte/opendraft/issues"

[project.scripts]
opendraft = "opendraft.cli:main"
draft-ai = "opendraft.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["opendraft*", "utils*", "concurrency*"]

[tool.setuptools]
py-modules = ["config", "draft_generator"]
include-package-data = true

[tool.setuptools.package-data]
"opendraft" = ["prompts/*.md", "prompts/**/*.md"]
"*" = ["*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --strict-markers --cov=. --cov-report=term-missing"
# Exclude standalone integration test scripts (run directly, not via pytest)
norecursedirs = ["tests/scripts", "tests/exploratory"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring API access",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
